pyproject.toml
[tool.poetry]
name = "cis245-testbank"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "cis245_testbank"}]

[tool.poetry.dependencies]
python = "^3.11"
black = "^23.10.1"
pylint = "^3.0.2"
regex = "^2023.10.3"
mysql-connector-python = "^8.2.0"
argon2-cffi = "^23.1.0"
pytest-mock = "^3.12.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
